Fix app test import boundaries
