+++ packages/app/src/components/ui/icon-color.ts
@@ -1,0 +1,9 @@
+import type { Theme } from "@/styles/theme";
+
+export const mutedIconColorMapping = (theme: Theme) => ({
+  color: theme.colors.foregroundMuted,
+});
+
+export const extraMutedIconColorMapping = (theme: Theme) => ({
+  color: theme.colors.foregroundExtraMuted,
+});
