+++ packages/app/src/components/ui/menu/index.ts
@@ -22,11 +22,11 @@ export {
   MenuLabel,
   MenuSeparator,
   MenuTextField,
-  MENU_ITEM_HEIGHT,
   menuRowContentInset,
   type ActionStatus,
   type MenuItemProps,
 } from "./menu-item";
+export { MENU_ITEM_HEIGHT } from "./menu-geometry";
 export {
   useMenuContext,
   useMenuDepth,
