+++ packages/app/src/components/ui/menu/menu-geometry.ts
@@ -1,0 +1,2 @@
+/** Shared geometry for menu rows and controls that align with them. */
+export const MENU_ITEM_HEIGHT = { xs: 40, md: 28 } as const;
